% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_populations.R
\name{prop_in_radius}
\alias{prop_in_radius}
\title{Calculate the proportion of each LSOA within a given radius of postcodes}
\usage{
prop_in_radius(df, radius_km, df_pc_col = "postcode")
}
\arguments{
\item{df}{A data frame containing the input postcodes and their
corresponding geographic coordinates. The data frame must contain a
postcode column and latitude and longitude columns.}

\item{radius_km}{Numeric. The radius around each input postcode, in
kilometres.}

\item{df_pc_col}{Character string. The name of the column in df
containing postcodes. Defaults to "postcode".}
}
\value{
A data frame containing one row for each LSOA represented by at
least one postcode within the specified radius, with the following
columns:
\describe{
\item{lsoa21_code}{The 2021 LSOA code.}
\item{pop_inside}{The estimated population within the specified
radius.}
\item{lsoa_pop}{The total estimated population of the LSOA.}
\item{pop_prop}{The estimated proportion of the LSOA population within
the specified radius.}
}
}
\description{
Estimates the proportion of the population of each Lower Layer Super
Output Area (LSOA) that is located within a specified radius of at least
one of a set of postcodes.
}
\details{
The function identifies postcodes within the specified radius of each
input postcode using the postcode coordinates in the package's
wm_postcodes dataset. Populations are aggregated to LSOA level, with
postcodes that fall within the radius of multiple input postcodes counted
only once. The returned proportion is therefore the estimated proportion
of the total LSOA population whose postcode centroid falls within the
specified radius of at least one input postcode.
}
