% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_distance.R
\name{nearest_postcode}
\alias{nearest_postcode}
\title{Find nearest postcode(s) between two datasets}
\usage{
nearest_postcode(
  df1,
  df2,
  df1_pc_col = "postcode",
  df2_pc_col = "postcode",
  n = 1,
  distance = FALSE
)
}
\arguments{
\item{df1}{A data frame containing input postcodes.}

\item{df2}{A data frame containing reference postcodes.}

\item{df1_pc_col}{Column name of postcodes in \code{df1}.}

\item{df2_pc_col}{Column name of postcodes in \code{df2}.}

\item{n}{Number of nearest neighbours to return.}

\item{distance}{Logical. If TRUE, returns distance between postcodes in df1
and nearest postcodes in df2}
}
\value{
A data frame with additional columns:
\itemize{
\item nearest_pc_1, nearest_pc_2, ... (Optionally) dist_km_1, dist_km_2,...
}
}
\description{
For each postcode in \code{df1}, finds the nearest postcode(s)
in \code{df2} using Euclidean distance on latitude/longitude.
}
\examples{
df1 <- data.frame(postcode = "B1 1BB")
df2 <- data.frame(postcode = c("B36 9ST", "B96 6DD", "B70 7EJ", "DY8 3YD"))
nearest_postcode(df1, df2, n = 2, distance = TRUE)
}
